Rename the browser tier's binary to yaak-web
It serves the web app and performs that app's requests, so it is named for
the tier rather than either half. `yaak serve` on the app binary stays free
for the headless tier that will hold real data.

The wasm crate takes the name it should have had: crates/yaak-web is the
code that runs inside the browser, so it becomes crates/yaak-wasm
(@yaakapp-internal/wasm), pkg/ regenerated. Env moves to YAAK_WEB_*, and
--serve-web loses its redundant half.
