+++ Cargo.toml
@@ -22,13 +22,13 @@ members = [
   "crates/yaak-templates",
   "crates/yaak-tls",
   "crates/yaak-ws",
-  "crates/yaak-web",
+  "crates/yaak-wasm",
   "crates/yaak-api",
   "crates/yaak-proxy",
   # Proxy-specific crates
   "crates-proxy/yaak-proxy-lib",
   # Server crates (the browser tier's hosted send executor)
-  "crates-server/yaak-server",
+  "crates-server/yaak-web",
   # CLI crates
   "crates-cli/yaak-cli",
   # Tauri-specific crates
