+++ crates-server/yaak-web/index.ts
@@ -1,4 +1,4 @@
 // The server's wire contract, generated by ts-rs from src/wire.rs
-// (`cargo test -p yaak-server`). The tab imports these so a change to a
+// (`cargo test -p yaak-web`). The tab imports these so a change to a
 // frame on the Rust side is a type error in packages/platform/src/web.
-export type { Frame, SendRequest } from "./bindings/gen_server";
+export type { Frame, SendRequest } from "./bindings/gen_web";
