+++ crates/yaak-models/src/models.rs
@@ -197,7 +197,7 @@ impl ResolvedHttpRequestSettings {
 
 /// The resolved send settings, values only: what an executor has to obey, with the sources
 /// (which model each came from) left behind in [`ResolvedHttpRequestSettings`]. This is what
-/// crosses from a tab to the send proxy, and what the proxy reads.
+/// crosses from a tab to the Yaak server, and what the server reads.
 #[derive(Debug, Clone, PartialEq, Serialize, Deserialize, TS)]
 #[serde(rename_all = "camelCase")]
 #[ts(export, export_to = "gen_models.ts")]
