+++ crates/yaak-wasm/Cargo.toml
@@ -1,5 +1,5 @@
 [package]
-name = "yaak-web"
+name = "yaak-wasm"
 version = "0.1.0"
 edition = "2024"
 publish = false
