+++ crates/yaak-wasm/package.json
@@ -1,5 +1,5 @@
 {
-  "name": "@yaakapp-internal/web",
+  "name": "@yaakapp-internal/wasm",
   "version": "1.0.0",
   "private": true,
   "main": "index.ts",
