+++ crates/yaak-wasm/src/lib.rs
