+++ packages/platform/package.json
@@ -11,8 +11,8 @@
   "dependencies": {
     "@yaakapp-internal/models": "^1.0.0",
     "@yaakapp-internal/rpc-schema": "^1.0.0",
-    "@yaakapp-internal/server": "^1.0.0",
     "@yaakapp-internal/web": "^1.0.0",
+    "@yaakapp-internal/wasm": "^1.0.0",
     "@tauri-apps/api": "^2.11.0",
     "@tauri-apps/plugin-clipboard-manager": "^2.3.2",
     "@tauri-apps/plugin-dialog": "^2.7.1",
