+++ packages/platform/src/web/server.ts
@@ -1,11 +1,11 @@
 /**
  * The wire to the Yaak server: where it is, and how to read what comes back.
  *
- * The shapes themselves are generated from `crates-server/yaak-server/src/wire.rs`
- * into `@yaakapp-internal/server`, so the two sides cannot drift silently.
+ * The shapes themselves are generated from `crates-server/yaak-web/src/wire.rs`
+ * into `@yaakapp-internal/web`, so the two sides cannot drift silently.
  */
 
-import type { Frame } from "@yaakapp-internal/server";
+import type { Frame } from "@yaakapp-internal/web";
 
 /* ------------------------------- location -------------------------------- */
 
@@ -13,17 +13,17 @@ import type { Frame } from "@yaakapp-internal/server";
  * Where the tab sends.
  *
  * Empty means "this origin": the server can serve the app itself
- * (`yaak-server --serve-web`), and then a send is a request to a path on the
+ * (`yaak-web --serve`), and then a send is a request to a path on the
  * page's own origin — no CORS, and nothing for a self-hoster to configure.
  *
- * `VITE_YAAK_SERVER_URL` overrides it at build time, for a deployment that
+ * `VITE_YAAK_WEB_URL` overrides it at build time, for a deployment that
  * keeps the two apart. The dev server is one of those: it serves the app on its
  * own origin and knows nothing about `/v1`, so a dev build falls back to a server
- * running locally (`cargo run -p yaak-server`).
+ * running locally (`cargo run -p yaak-web`).
  */
 export function serverBaseUrl(): string {
   const env = (import.meta as unknown as { env?: Record<string, string | undefined> }).env;
-  const configured = env?.VITE_YAAK_SERVER_URL?.trim();
+  const configured = env?.VITE_YAAK_WEB_URL?.trim();
   if (configured) return configured.replace(/\/+$/, "");
   return env?.DEV ? "http://127.0.0.1:9227" : "";
 }
@@ -40,14 +40,14 @@ function serverLocation(): string {
 }
 
 /**
- * Who does the sending, for the timeline: `yaak-server 0.1.0 at http://…`.
+ * Who does the sending, for the timeline: `yaak-web 0.1.0 at http://…`.
  * Asked of `/v1/health` once per page load; if the server can't be reached the
  * URL alone is the answer, and the send itself will say why shortly after.
  */
 export function serverIdentity(): Promise<string> {
   identity ??= fetch(`${serverBaseUrl()}/v1/health`)
     .then((res) => res.json() as Promise<{ version?: string }>)
-    .then((health) => `yaak-server ${health.version ?? ""} at ${serverLocation()}`.replace("  ", " "))
+    .then((health) => `yaak-web ${health.version ?? ""} at ${serverLocation()}`.replace("  ", " "))
     .catch(() => {
       identity = null; // try again next send
       return `Yaak server at ${serverLocation()}`;
